backend/requirements.txt
# BTC Watcher Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# WebSocket
websockets==12.0

# Task Queue & Background Jobs
celery==5.3.4

# Monitoring & Logging
psutil==5.9.6
python-json-logger==2.0.7

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Data Processing
pandas==2.1.4
numpy==1.26.2

# FreqTrade (will be installed in separate container)
# freqtrade==2025.8

# Utilities
aiofiles==23.2.1
pyyaml==6.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
